Attempt to make openssl optional by splitting packages and using features
