+++ Cargo.toml
@@ -9,6 +9,39 @@ members = [
   "crates/*",
 ]
 
+[workspace.dependencies]
+ndc-sdk-core = { path = "../sdk-core" }
+ndc-models = { git = "http://github.com/hasura/ndc-spec.git", tag = "v0.1.6" }
+ndc-test = { git = "http://github.com/hasura/ndc-spec.git", tag = "v0.1.6" }
+
+anyhow = "1"
+async-trait = "0.1"
+axum = { version = "0.6", features = ["http2"] }
+axum-extra = "0.8"
+bytes = "1"
+clap = { version = "4", features = ["derive", "env"] }
+http = "0.2"
+mime = "0.3"
+opentelemetry = "0.22"
+opentelemetry-http = "0.11"
+opentelemetry-otlp = { version = "0.15", features = ["reqwest-client", "gzip-tonic", "tls", "tls-roots", "http-proto"] }
+opentelemetry-semantic-conventions = "0.14"
+opentelemetry_sdk = { version = "0.22", features = ["rt-tokio"] }
+opentelemetry-zipkin = "0.20"
+prometheus = "0.13"
+reqwest = "0.11"
+serde = { version = "1", features = ["derive"] }
+serde_json = { version = "1", features = ["raw_value"] }
+thiserror = "1"
+tokio = { version = "1", features = ["fs", "macros", "rt-multi-thread", "signal"] }
+tokio-test = "0.4"
+tower-http = { version = "0.4", features = ["cors", "limit", "trace", "validate-request"] }
+tracing = "0.1"
+tracing-opentelemetry = "0.23"
+tracing-subscriber = { version = "0.3", default-features = false, features = ["ansi", "env-filter", "fmt", "json"] }
+url = "2"
+
+
 [workspace.lints.clippy]
 all = { level = "warn", priority = -1 }
 pedantic = { level = "warn", priority = -1 }
