+++ crates/sdk-core/Cargo.toml
@@ -1,0 +1,41 @@
+[package]
+name = "ndc-sdk-core"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+
+[lints]
+workspace = true
+
+[lib]
+name = "ndc_sdk_core"
+path = "src/lib.rs"
+
+[features]
+default = ["axum", "ndc-test"]
+
+axum = ["dep:axum", "dep:mime"]
+
+ndc-test = ["dep:ndc-test"]
+
+[dependencies]
+ndc-models = { workspace = true }
+ndc-test = { workspace = true, optional = true }
+
+async-trait = { workspace = true }
+axum = { workspace = true, features = ["http2"], optional = true }
+bytes = { workspace = true }
+http = { workspace = true }
+mime = { workspace = true, optional = true }
+prometheus = { workspace = true }
+serde = { workspace = true, features = ["derive"] }
+serde_json = { workspace = true, features = ["raw_value"] }
+thiserror = { workspace = true }
+tokio = { workspace = true, features = ["fs", "macros", "rt-multi-thread", "signal", "sync"] }
+tracing = { workspace = true }
+
+[dev-dependencies]
+anyhow = { workspace = true }
+axum = { workspace = true, features = ["http2"] }
+reqwest = { workspace = true }
+tokio-test = { workspace = true }
