+++ crates/sdk-core/src/connector/example.rs
