+++ crates/sdk/src/lib.rs
@@ -1,10 +1,10 @@
 pub mod check_health;
-pub mod connector;
 pub mod default_main;
 pub mod fetch_metrics;
 pub mod json_rejection;
-pub mod json_response;
-pub mod state;
 pub mod tracing;
 
 pub use ndc_models as models;
+pub use ndc_sdk_core::connector;
+pub use ndc_sdk_core::json_response;
+pub use ndc_sdk_core::state;
